[Refactor] Search, DetailPost, Login, SignUp 페이지 컴포넌트 분리
