+++ .env
@@ -1,5 +1,6 @@
 REACT_APP_KAKAO_REST_API_KEY=a558b560602bb0e31eec3dcf9866557e
 REACT_APP_KAKAO_REST_SECRET=u8rQ5HJHR7WhNfSWqe7HJpqz2Hau4Iax
 REACT_APP_KAKAO_REDIRECT_URI=http://localhost:3000/auth/kakao/redirect
-REACT_APP_SERVER_URL=http://3.35.19.8:3000/api
+# REACT_APP_SERVER_URL=http://3.35.19.8:3000/api
+REACT_APP_SERVER_URL=http://13.209.50.102/api
 REACT_APP_SOCKET_URL=http://52.78.166.176:3000
