+++ src/components/elem/Icons.jsx
@@ -1,0 +1,18 @@
+const Icons = [
+  "👗",
+  "🍱",
+  "💰",
+  "💘",
+  "🎨",
+  "⚽️",
+  "✈️",
+  "👩‍❤️‍👨",
+  "🎮",
+  "🐶",
+  "👨‍👩‍👧‍👦",
+  "💼",
+  "💬",
+  "🎸",
+];
+
+export default Icons;
