+++ src/components/header/LoginHeader.jsx
@@ -1,0 +1,18 @@
+import React from "react";
+import * as St from "./LoginHeader.style";
+import { useNavigate } from "react-router-dom";
+
+const LoginHeader = () => {
+  const navi = useNavigate();
+  return (
+    <St.Box
+      onClick={() => {
+        navi("/");
+      }}
+    >
+      <St.Logo />
+    </St.Box>
+  );
+};
+
+export default LoginHeader;
