+++ src/components/post/NewPostCardSlider.jsx
@@ -1,10 +1,7 @@
 import React, { useEffect, useState } from "react";
-import styled, { css } from "styled-components";
-import { instance } from "../../api/axios";
-import sliderBtnLeft from "../../assets/icons/common/sliderBtnLeft.png";
 import NewPostCard from "./NewPostCard";
-import axios from "axios";
-import { useNavigate } from "react-router-dom";
+import { instance } from "../../api/axios";
+import * as St from "./NewPostCardSlider.style";
 
 const NewPostCardSlider = () => {
   // newPostCard 불러오기
@@ -65,12 +62,12 @@ const NewPostCardSlider = () => {
   };
 
   return (
-    <XXXNewPostCardSliderCont>
-      <XXXPostCardSliderInfo title="실시간 훈수" more="더 보기" />
-      <SliderContainer>
-        <Button dir="left" onClick={clickLeftHandler} />
-        <Button dir="right" onClick={clickRightHandler} />
-        <NewPostCards animate={animate}>
+    <St.Wrap>
+      <St.Info title="실시간 훈수" more="더 보기" />
+      <St.Cont>
+        <St.Btn dir="left" onClick={clickLeftHandler} />
+        <St.Btn dir="right" onClick={clickRightHandler} />
+        <St.Cards animate={animate}>
           {genPostCardsArray(index)?.map((el) => (
             <NewPostCard
               key={el?.postIdx}
@@ -81,148 +78,10 @@ const NewPostCardSlider = () => {
               content={el?.desc}
             />
           ))}
-        </NewPostCards>
-      </SliderContainer>
-    </XXXNewPostCardSliderCont>
+        </St.Cards>
+      </St.Cont>
+    </St.Wrap>
   );
 };
 
 export default NewPostCardSlider;
-
-const XXXNewPostCardSliderCont = styled.li``;
-
-const XXXPostCardSliderInfo = ({ colored, title, more, on }) => {
-  const nav = useNavigate();
-  return (
-    <XXXPostCardSliderInfoXXX>
-      <XXXPostCardSliderTitle>
-        <span className="colored">{colored}</span>
-        {title}
-      </XXXPostCardSliderTitle>
-      <XXXPostCardSliderMore
-        onClick={() => {
-          if (on !== "on") {
-            nav("/totalboard");
-          } else {
-            nav("/battle");
-          }
-        }}
-      >
-        {more}
-      </XXXPostCardSliderMore>
-    </XXXPostCardSliderInfoXXX>
-  );
-};
-
-const XXXPostCardSliderInfoXXX = styled.div`
-  /* border: 1px solid olivedrab; */
-  width: 350px;
-  height: 26px;
-  margin-left: 25px;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const XXXPostCardSliderTitle = styled.h1`
-  /* border: 1px solid red; */
-  font-weight: 600;
-  font-size: 20px;
-  color: #2d2d2d;
-
-  /* colored에 스타일주기 */
-  span.colored {
-    color: #ef3f61;
-  }
-`;
-
-const XXXPostCardSliderMore = styled.span`
-  /* border: 1px solid blue; */
-  color: #ef3f61;
-  font-size: 0.8rem;
-  cursor: pointer;
-`;
-
-const SliderContainer = styled.div`
-  /* border: 1px solid green; */
-  width: 400px;
-  height: 160px;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  position: relative;
-  overflow-x: hidden;
-`;
-
-const NewPostCards = styled.div`
-  display: flex;
-
-  gap: 8px;
-
-  ${({ animate }) => {
-    if (animate.on) {
-      return css`
-        transform: translate(${({ animate }) => animate.value});
-        transition: transform 350ms ease-in-out;
-      `;
-    }
-  }};
-
-  /* & :hover {
-    // 내부 컨텐츠들은 안 움직였으면 좋겠는데....
-    transform: scale(1.05);
-  } */
-`;
-
-// 버튼
-const Button = ({ dir, onClick }) => {
-  return (
-    <Stbutton dir={dir} onClick={onClick}>
-      <ButtonImg src={sliderBtnLeft} />
-    </Stbutton>
-  );
-};
-
-const Stbutton = styled.button`
-  background-color: transparent;
-  width: 20px;
-  height: 20px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border: none;
-  border-radius: 50%;
-  position: absolute;
-  z-index: 1;
-  cursor: pointer;
-  background: rgba(46, 46, 71, 0.4);
-
-  ${({ dir }) => {
-    if (dir === "left") {
-      return css`
-        left: 25px;
-        top: 50%;
-        transform: translateY(-50%);
-      `;
-    }
-
-    if (dir === "right") {
-      return css`
-        right: 25px;
-        top: 50%;
-        transform: translateY(-50%) scaleX(-1);
-      `;
-    }
-  }}
-`;
-
-const ButtonImg = styled.img`
-  width: 5px;
-  height: 10px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
